// simulation/types.ts
export type GovernorType = 'oz' | 'bravo'

interface SimulationConfigBase {
  type: 'executed' | 'proposed' | 'new'
  daoName: string // e.g. 'Compound' or 'Uniswap'
  governorAddress: string // address of the governor
  governorType: GovernorType
}

export interface SimulationConfigExecuted extends SimulationConfigBase {
  type: 'executed'
  proposalId: BigNumberish // ID of the executed proposal
}

export interface SimulationConfigProposed extends SimulationConfigBase {
  type: 'proposed'
  proposalId: BigNumberish // ID of the executed proposal
}

export interface SimulationConfigNew extends SimulationConfigBase {
  type: 'new'
  targets: string[]
  values: BigNumberish[]
  signatures: string[]
  calldatas: string[]
  description: string
}

export type SimulationConfig = SimulationConfigExecuted | SimulationConfigProposed | SimulationConfigNew

export interface SimulationResult {
  sim: TenderlySimulation
  proposal: ProposalEvent
  latestBlock: Block
}

export interface SimulationData extends SimulationResult {
  config: SimulationConfig
}
// ...
export interface ProposalEvent {
  id?: BigNumber // Bravo governor
  proposalId?: BigNumber // OZ governor
  proposer: string
  startBlock: BigNumber
  endBlock: BigNumber
  description: string
  targets: string[]
  values: BigNumber[]
  signatures: string[]
  calldatas: string[]
}

export type Message = string

export type CheckResult = {
  info: Message[]
  warnings: Message[]
  errors: Message[]
}

export type ProposalData = {
  governor: Contract
  timelock: Contract
  provider: JsonRpcProvider
}

export interface ProposalCheck {
  name: string
  checkProposal(proposal: ProposalEvent, tx: TenderlySimulation, deps: ProposalData): Promise<CheckResult>
}

export interface AllCheckResults {
  [checkId: string]: { name: string; result: CheckResult }
}
// ...
export interface TenderlySimulation {
  transaction: Transaction
  simulation: Simulation
  contracts: TenderlyContract[]
  generated_access_list: GeneratedAccessList[]
}